+++ src/Enums/ToastDirection.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Victorycodedev\ToastKit\Enums;
+
+enum ToastDirection: string
+{
+    case Auto = 'auto';
+    case Left = 'left';
+    case Right = 'right';
+    case Top = 'top';
+    case Bottom = 'bottom';
+}