+++ src/Exceptions/InvalidToastConfigurationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Victorycodedev\ToastKit\Exceptions;
+
+class InvalidToastConfigurationException extends ToastKitException {}