+++ src/Exceptions/PresetNotFoundException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Victorycodedev\ToastKit\Exceptions;
+
+class PresetNotFoundException extends ToastKitException {}